api/requirements.txt
fastapi
uvicorn
python-dotenv
browser-use
pydantic
requests
supabase
python-jose